Generalize the way to check if an Arm is operational to Bodyparts
