+++ robot_skills/src/robot_skills/robot_part.py
@@ -4,6 +4,7 @@
 import rospy
 import actionlib
 
+from diagnostic_msgs.msg import DiagnosticArray
 
 class RobotPart(object):
     """ Base class for robot parts """
@@ -18,8 +19,13 @@ class RobotPart(object):
 
         self.__ros_connections = {}
 
+        self.__diagnostics_name = ""
+        self._hardware_status_sub = None
 
+        # This is set to False by start_check_operational, because then apparently there is a meaningful check
+        # If no such check exists, then assume it's operational unless overridden in subclass
         self._operational = True
+
     def load_param(self, param_name, default=None):
         """
         Loads a parameter from the parameter server, namespaced by robot name
@@ -126,5 +132,41 @@ class RobotPart(object):
         """
         return self._operational
 
+    def start_check_operational(self, name):
+        """
+        Start to check if the bodypart is operational. To do so, subscribe to the hardware status/diagnostics
 
-        raise NotImplementedError("Implement in subclasses")
+        :param name: check on the level of the diagnostic_msgs/DiagnosticStatus with this name
+        """
+
+        self._operational = False
+
+        self.__diagnostics_name = name
+
+        self._hardware_status_sub = self.create_subscriber("/" + self.robot_name + "/hardware_status",
+                                                           DiagnosticArray, self.cb_hardware_status)
+
+    def cb_hardware_status(self, msg):
+        """
+        hardware_status callback to determine if the bodypart is operational
+        :param msg: diagnostic_msgs.msg.DiagnosticArray
+        :return: no return
+        """
+        diags = [diag for diag in msg.status if diag.name == self.__diagnostics_name]
+
+        if len(diags) == 0:
+            rospy.logwarn('no diagnostic msg received for the %s' % self.__diagnostics_name)
+        elif len(diags) != 1:
+            rospy.logwarn('multiple diagnostic msgs received for the %s' % self.__diagnostics_name)
+        else:
+            level = diags[0].level
+
+            # 0. Stale
+            # 1. Idle
+            # 2. Operational
+            # 3. Homing
+            # 4. Error
+            if level != 2:
+                self._operational = False
+            else:
+                self._operational = True
